// packages/shc-types/src/schemas/cook.d.ts
import { z } from 'zod';
export declare const CookStatus: z.ZodEnum<["pending", "active", "paused", "suspended"]>;
export type CookStatus = z.infer<typeof CookStatus>;
export declare const shcCookSchema: z.ZodObject<{
    id: z.ZodString;
    auth_identity_id: z.ZodString;
    slug: z.ZodString;
    display_name: z.ZodString;
    story: z.ZodOptional<z.ZodString>;
    area: z.ZodString;
    collection_address: z.ZodOptional<z.ZodString>;
    collection_instructions: z.ZodOptional<z.ZodString>;
    status: z.ZodEnum<["pending", "active", "paused", "suspended"]>;
    availability_paused: z.ZodDefault<z.ZodBoolean>;
    expo_push_token: z.ZodOptional<z.ZodString>;
    sfa_reg_number: z.ZodOptional<z.ZodString>;
    wsq_cert_expiry: z.ZodOptional<z.ZodString>;
    pdpa_consent_at: z.ZodOptional<z.ZodString>;
    pdpa_consent_version: z.ZodOptional<z.ZodString>;
    created_at: z.ZodOptional<z.ZodString>;
    updated_at: z.ZodOptional<z.ZodString>;
}, "strict", z.ZodTypeAny, {
    status: "pending" | "active" | "paused" | "suspended";
    id: string;
    auth_identity_id: string;
    slug: string;
    display_name: string;
    area: string;
    availability_paused: boolean;
    story?: string | undefined;
    collection_address?: string | undefined;
    collection_instructions?: string | undefined;
    expo_push_token?: string | undefined;
    sfa_reg_number?: string | undefined;
    wsq_cert_expiry?: string | undefined;
    pdpa_consent_at?: string | undefined;
    pdpa_consent_version?: string | undefined;
    created_at?: string | undefined;
    updated_at?: string | undefined;
}, {
    status: "pending" | "active" | "paused" | "suspended";
    id: string;
    auth_identity_id: string;
    slug: string;
    display_name: string;
    area: string;
    story?: string | undefined;
    collection_address?: string | undefined;
    collection_instructions?: string | undefined;
    availability_paused?: boolean | undefined;
    expo_push_token?: string | undefined;
    sfa_reg_number?: string | undefined;
    wsq_cert_expiry?: string | undefined;
    pdpa_consent_at?: string | undefined;
    pdpa_consent_version?: string | undefined;
    created_at?: string | undefined;
    updated_at?: string | undefined;
}>;
export type SHCCook = z.infer<typeof shcCookSchema>;
//# sourceMappingURL=cook.d.ts.map